requirements.txt
watchdog
python-dotenv
aiogram